requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
pdf2image==1.16.0
Pillow==8.3.1
pytesseract==0.3.8
